cpu-pool-policy: move to int64 for capacity.

+++ pkg/apis/cpupools.intel.com/v1alpha1/types.go
@@ -4,8 +4,8 @@ import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 type Pool struct {
 	PoolName  string `json:"poolName"`
-	Usage     int    `json:"usage"`
-	Capacity  int    `json:"capacity"`
+	Usage     int64  `json:"usage"`
+	Capacity  int64  `json:"capacity"`
 	Exclusive string `json:"exclusive"`
 	Shared    string `json:"shared"`
 }
